requirements.txt
# Cold Email Campaign Orchestrator
anthropic>=0.40.0
streamlit>=1.30.0
requests>=2.31.0
duckduckgo-search>=6.0.0
flask>=3.0.0
gspread>=6.0.0
google-auth>=2.25.0
python-dotenv>=1.0.0
python-whois>=0.9.4
tavily-python>=0.5.0
